Refactor out assignment logic into service objects
